requirements.txt
streamlit
finance-datareader
google-generativeai
pandas